Seetaram: Refactoring Game and Player.

+++ .eslintrc.js
@@ -6,7 +6,7 @@ module.exports = {
   "parser": "babel-eslint",
   "rules": {
     "semi": ["error", "always"],
-    "quotes": ["error", "double"]
+    "quotes": ["error", "single"]
   }
 };
